update risk_score to decimal

+++ server/database.sql
@@ -17,7 +17,7 @@ create table driver (
     name varchar(100) not null,
     active boolean default false,
     phone_number varchar(15) unique not null,
-    risk_score integer default 0,
+    risk_score decimal(5,2) default 0,
     fleet_id integer not null,
     foreign key (fleet_id) references fleet(id) on delete cascade
 );
